some stuff

+++ app/src/main/java/com/example/project2/Activities/ShipViewerActivity.java
@@ -17,15 +17,14 @@ import com.example.project2.DB.FleetDAO;
 import com.example.project2.DB.FleetsTableDAO;
 import com.example.project2.DB.StarShipDAO;
 import com.example.project2.DialogJunk.ShipDeleteConfirmationDialog;
-import com.example.project2.Observer.Observer;
 import com.example.project2.Observer.PositionObserver;
 import com.example.project2.R;
 import com.example.project2.RecycleViewStuff.ShipRecyclerAdapter;
-import com.example.project2.RecycleViewStuff.UserAdapterView;
 import com.example.project2.StarConfData.Fleet;
 import com.example.project2.StarConfData.Ship;
 import com.example.project2.databinding.ActivityShipViewerBinding;
 
+import java.util.ArrayList;
 import java.util.List;
 
 /*FINISHED ACTIVITY
@@ -78,13 +77,9 @@ public class ShipViewerActivity extends AppCompatActivity implements
 
         mShipList = mStarShipDAO.getAllShips();
 
-        mRecyclerView = findViewById(R.id.recyclerView);
-        mRecyclerView.setHasFixedSize(true);
-        mLayoutManager = new LinearLayoutManager(this);
-        mAdapter = new ShipRecyclerAdapter(mShipList);
 
-        mRecyclerView.setLayoutManager(mLayoutManager);
-        mRecyclerView.setAdapter(mAdapter);
+
+        refreshRecyclerView();
 
         mAdapter.setOnItemClickListener(new ShipRecyclerAdapter.OnItemClickListener() {
             @Override
@@ -102,10 +97,15 @@ public class ShipViewerActivity extends AppCompatActivity implements
 
     @Override
     public void onYesClicked(){
-        List<Fleet> fleetsToDelete = mFleetsTableDAO.getFleetsByShipType(mShipList.get(observer.getPosition()).getShipLogId());
+        List<Fleet> fleetsToDelete = mFleetDAO.getFleets();
         for(Fleet fleet : fleetsToDelete){
-            mFleetDAO.delete(fleet);
+            if(fleet.getFleet().contains(mShipList.get(observer.getPosition()).getShipLogId())){
+                mFleetDAO.delete(fleet);
+            }
         }
+        mShipList.remove(observer.getPosition());
+        mAdapter.notifyItemRemoved(observer.getPosition());
+        mStarShipDAO.delete(mShipList.get(observer.getPosition()));
     }
 
     @Override
@@ -117,7 +117,9 @@ public class ShipViewerActivity extends AppCompatActivity implements
         item.setOnMenuItemClickListener(new MenuItem.OnMenuItemClickListener() {
             @Override
             public boolean onMenuItemClick(@NonNull MenuItem menuItem) {
-                return false;
+                Intent intent = AdministratorPageActivity.intentFactory(getApplicationContext(), true, getIntent().getStringExtra(MESSAGE));
+                startActivity(intent);
+                return true;
             }
         });
         return super.onCreateOptionsMenu(menu);
@@ -127,5 +129,18 @@ public class ShipViewerActivity extends AppCompatActivity implements
         Intent intent = new Intent(context, ShipViewerActivity.class);
         intent.putExtra(MESSAGE, loggedUser);
         return intent;
+    }
+
+    public void refreshRecyclerView(){
+        mRecyclerView = findViewById(R.id.fleetRecyclerView);
+        mLayoutManager = new LinearLayoutManager(this);
+        ArrayList<Ship> tempList = new ArrayList<>();
+        for(Ship ship : mShipList){
+            tempList.add(ship);
+        }
+        mAdapter = new ShipRecyclerAdapter(tempList);
+
+        mRecyclerView.setLayoutManager(mLayoutManager);
+        mRecyclerView.setAdapter(mAdapter);
     }
 }
